fix(donations): replace MySQL-only ST_Distance_Sphere with portable Haversine
scopeNearLocation was using ST_Distance_Sphere(POINT(...)) which only exists
in MySQL. The project runs SQLite in dev and PostgreSQL in production.

Replace with a two-step approach that works on any database:
1. SQL bounding box (whereBetween lat/lon) filters candidates without any
   math extension dependency.
2. Exact Haversine calculation in PHP sorts and filters the bounding box
   results, guaranteeing identical behavior in SQLite and PostgreSQL.

Also expose latitude/longitude/radius as GET /donation-items query params
and add DonationItemFactory with a withCoordinates() helper for tests.

Refs: TCK-002

+++ app/Contracts/Services/DonationItemServiceInterface.php
@@ -41,7 +41,7 @@ interface DonationItemServiceInterface
     /**
      * Buscar itens por localização
      */
-    public function findByLocation(float $latitude, float $longitude, int $radius = 10, int $perPage = 15): LengthAwarePaginator;
+    public function findByLocation(float $latitude, float $longitude, float $radius = 10, int $perPage = 15): LengthAwarePaginator;
 
     /**
      * Marcar item como doado
@@ -58,4 +58,3 @@ interface DonationItemServiceInterface
      */
     public function canUserModify(DonationItem $item, User $user): bool;
 }
-
